Add the event loop as required constructor argument
For now we won't use it just yet. But this will prepare our API for
future features like keep alive connections, dead connection clean up
etc. Even though we won't use it in the next release, it will put it in
the API so we don't have to break it again when we add features that
require the event loop.

As a result of this requirement the event loop has been added as a
dependency, and the socket and stream packages have been bumped to
their stable version.