Require Host request header for HTTP/1.1 requests
All HTTP/1.1 requests require a Host request header as per RFC 7230. The
proxy CONNECT method is an exception to this rule because we do not want
to break compatibility with common HTTP proxy clients that do not
strictly follow the RFCs.

This does not affect valid HTTP/1.1 requests and has no effect on
HTTP/1.0 requests.

Additionally, make sure we do not include a default Host request header
in the parsed request object if the incoming request does not make use
of the Host request header.